requirements.txt
django==1.4
django-celery
django-tastypie
setproctitle
oauth2